Experimenting with 32 bits positions

+++ src/temp.c
@@ -4,5 +4,5 @@ unsigned char arg2;
 unsigned char pad1;
 unsigned char pad1_new;
 unsigned char temp, i, j, k, temp_x, temp_y, temp_char, temp_attr, temp_bank;
-unsigned int temp_int, temp_int_x, temp_int_y;
+unsigned long temp_int, temp_int_x, temp_int_y;
 #pragma bss-name(pop)
